+++ src/app/static/src/app/components/modelOutput/ModelOutput.vue
@@ -47,6 +47,7 @@
                     :indicators="barchartIndicators"
                     :selections="barchartSelections"
                     :formatFunction="formatBarchartValue"
+                    :showRangesInTooltips="true"
                     @update="updateBarchartSelectionsAndXAxisOrder"></bar-chart-with-filters>
                 <div class="row mt-2">
                     <div class="col-md-3"></div>